day124/day124-azure-monitor-integration/requirements.txt
# Azure SDK
azure-identity==1.16.1
azure-monitor-query==1.4.1
azure-mgmt-monitor==6.0.2

# Web Framework
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0
jinja2==3.1.4

# Data Processing
aiohttp==3.9.5
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
pydantic==2.7.1
pytest==8.2.2
pytest-asyncio==0.23.7

# Frontend
requests==2.31.0
